Move dir history into terminalContrib/
